fix: move atoms to store
